fix: refactor to use function
